requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.0
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
aiohappyeyeballs=2.4.3=pyhd8ed1ab_0
aiohttp=3.10.10=py310h89163eb_0
aiosignal=1.3.1=pyhd8ed1ab_0
alsa-lib=1.2.13=hb9d3cd8_0
async-timeout=4.0.3=pyhd8ed1ab_0
attrs=24.2.0=pyh71513ae_0
aws-c-auth=0.8.0=hcd8ed7f_7
aws-c-cal=0.8.0=he70792b_1
aws-c-common=0.10.0=hb9d3cd8_0
aws-c-compression=0.3.0=hba2fe39_1
aws-c-event-stream=0.5.0=h127f702_4
aws-c-http=0.9.0=h8a7d7e2_5
aws-c-io=0.15.1=h2a50c78_1
aws-c-mqtt=0.11.0=hd25e75f_5
aws-c-s3=0.7.0=h858c4ad_7
aws-c-sdkutils=0.2.1=hba2fe39_0
aws-checksums=0.2.0=hba2fe39_1
aws-crt-cpp=0.29.3=hbc793f2_2
aws-sdk-cpp=1.11.407=h5cd358a_9
azure-core-cpp=1.14.0=h5cfcd09_0
azure-identity-cpp=1.10.0=h113e628_0
azure-storage-blobs-cpp=12.13.0=h3cf044e_1
azure-storage-common-cpp=12.8.0=h736e048_1
azure-storage-files-datalake-cpp=12.12.0=ha633028_1
blas=2.116=mkl
blas-devel=3.9.0=16_linux64_mkl
brotli=1.1.0=hb9d3cd8_2
brotli-bin=1.1.0=hb9d3cd8_2
brotli-python=1.1.0=py310hf71b8c6_2
bzip2=1.0.8=h4bc722e_7
c-ares=1.34.3=heb4867d_0
ca-certificates=2024.8.30=hbcca054_0
cairo=1.18.0=hebfffa5_3
certifi=2024.8.30=pyhd8ed1ab_0
cffi=1.17.1=py310h8deb56e_0
charset-normalizer=3.4.0=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
conllu=6.0.0=pyhd8ed1ab_0
contourpy=1.3.1=py310h3788b33_0
cpython=3.10.15=py310hd8ed1ab_2
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cuda-version=12.6=3
cycler=0.12.1=pyhd8ed1ab_0
cyrus-sasl=2.1.27=h54b06d7_7
datasets=3.0.2=pyhd8ed1ab_0
dbus=1.13.6=h5008d03_3
dill=0.3.8=pyhd8ed1ab_0
double-conversion=3.3.0=h59595ed_0
expat=2.6.4=h5888daf_0
filelock=3.16.1=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_3
fontconfig=2.15.0=h7e30c49_1
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.54.1=py310h89163eb_1
freetype=2.12.1=h267a509_2
frozenlist=1.5.0=py310ha75aee5_0
fsspec=2024.9.0=pyhff2d567_0
gflags=2.2.2=h5888daf_1005
glog=0.7.1=hbabe93e_0
gmp=6.3.0=hac33072_2
gmpy2=2.1.5=py310he8512ff_2
graphite2=1.3.13=h59595ed_1003
h2=4.1.0=pyhd8ed1ab_0
harfbuzz=9.0.0=hda332d3_1
hmmlearn=0.3.3=py310hf71b8c6_0
hpack=4.0.0=pyh9f0ad1d_0
huggingface_hub=0.26.2=pyh0610db2_0
hyperframe=6.0.1=pyhd8ed1ab_0
icu=75.1=he02047a_0
idna=3.10=pyhd8ed1ab_0
jinja2=3.1.4=pyhd8ed1ab_0
joblib=1.4.2=pyhd8ed1ab_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.7=py310h3788b33_0
krb5=1.21.3=h659f571_0
lcms2=2.16=hb7c19ff_0
ld_impl_linux-64=2.43=h712a8e2_2
lerc=4.0.0=h27087fc_0
libabseil=20240722.0=cxx17_h5888daf_1
libarrow=18.0.0=h3e543c6_4_cpu
libarrow-acero=18.0.0=h5888daf_4_cpu
libarrow-dataset=18.0.0=h5888daf_4_cpu
libarrow-substrait=18.0.0=h5c8f2c3_4_cpu
libblas=3.9.0=16_linux64_mkl
libbrotlicommon=1.1.0=hb9d3cd8_2
libbrotlidec=1.1.0=hb9d3cd8_2
libbrotlienc=1.1.0=hb9d3cd8_2
libcblas=3.9.0=16_linux64_mkl
libclang-cpp19.1=19.1.3=default_hb5137d0_0
libclang13=19.1.3=default_h9c6a7e4_0
libcrc32c=1.1.2=h9c3ff4c_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.11.1.6=0
libcups=2.3.3=h4637d8d_4
libcurand=10.3.7.77=0
libcurl=8.10.1=hbbe4b11_0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.22=hb9d3cd8_0
libdrm=2.4.123=hb9d3cd8_0
libedit=3.1.20191231=he28a2e2_2
libegl=1.7.0=ha4b6fd6_2
libev=4.33=hd590300_2
libevent=2.1.12=hf998b51_1
libexpat=2.6.4=h5888daf_0
libffi=3.4.2=h7f98852_5
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgfortran=14.2.0=h69a702a_1
libgfortran-ng=14.2.0=h69a702a_1
libgfortran5=14.2.0=hd5240d6_1
libgl=1.7.0=ha4b6fd6_2
libglib=2.82.2=h2ff4ddf_0
libglvnd=1.7.0=ha4b6fd6_2
libglx=1.7.0=ha4b6fd6_2
libgomp=14.2.0=h77fa898_1
libgoogle-cloud=2.31.0=h804f50b_0
libgoogle-cloud-storage=2.31.0=h0121fbd_0
libgrpc=1.67.1=hc2c308b_0
libhwloc=2.11.1=default_hecaa2ac_1000
libiconv=1.17=hd590300_2
libjpeg-turbo=3.0.0=hd590300_1
liblapack=3.9.0=16_linux64_mkl
liblapacke=3.9.0=16_linux64_mkl
libllvm19=19.1.3=ha7bfdaf_0
libnghttp2=1.64.0=h161d5f1_0
libnpp=11.8.0.86=0
libnsl=2.0.1=hd590300_0
libntlm=1.4=h7f98852_1002
libnvjpeg=11.9.0.86=0
libopengl=1.7.0=ha4b6fd6_2
libparquet=18.0.0=h6bd9018_4_cpu
libpciaccess=0.18=hd590300_0
libpng=1.6.44=hadc24fc_0
libpq=17.1=h04577a9_0
libprotobuf=5.28.2=h5b01275_0
libre2-11=2024.07.02=hbbce691_1
libsqlite=3.47.0=hadc24fc_1
libssh2=1.11.0=h0841786_0
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-ng=14.2.0=h4852527_1
libthrift=0.21.0=h0e7cc3e_0
libtiff=4.7.0=he137b08_1
libutf8proc=2.8.0=h166bdaf_0
libuuid=2.38.1=h0b41bf4_0
libwebp-base=1.4.0=hd590300_0
libxcb=1.17.0=h8a09558_0
libxcrypt=4.4.36=hd590300_1
libxkbcommon=1.7.0=h2c5496b_1
libxml2=2.13.5=hb346dea_0
libxslt=1.1.39=h76b75d6_0
libzlib=1.3.1=hb9d3cd8_2
llvm-openmp=15.0.7=h0cdce71_0
lz4-c=1.9.4=hcb278e6_0
markupsafe=3.0.2=py310h89163eb_0
matplotlib=3.9.2=py310hff52083_2
matplotlib-base=3.9.2=py310h68603db_2
mkl=2022.1.0=h84fe81f_915
mkl-devel=2022.1.0=ha770c72_916
mkl-include=2022.1.0=h84fe81f_915
mpc=1.3.1=h24ddda3_1
mpfr=4.2.1=h90cbb55_3
mpmath=1.3.0=pyhd8ed1ab_0
multidict=6.1.0=py310h89163eb_1
multiprocess=0.70.16=py310ha75aee5_1
munkres=1.1.4=pyh9f0ad1d_0
mysql-common=9.0.1=h266115a_2
mysql-libs=9.0.1=he0572af_2
ncurses=6.5=he02047a_1
networkx=3.4.2=pyh267e887_2
numpy=2.1.3=py310hd6e36ab_0
openjpeg=2.5.2=h488ebb8_0
openldap=2.6.8=hedd0468_0
openssl=3.4.0=hb9d3cd8_0
orc=2.0.2=he039a57_2
packaging=24.2=pyhd8ed1ab_0
pandas=2.2.3=py310h5eaa309_1
pcre2=10.44=hba22ea6_2
pillow=11.0.0=py310hfeaa1f3_0
pip=24.3.1=pyh8b19718_0
pixman=0.43.2=h59595ed_0
propcache=0.2.0=py310ha75aee5_2
pthread-stubs=0.4=hb9d3cd8_1002
pyarrow=18.0.0=py310hff52083_1
pyarrow-core=18.0.0=py310hac404ae_1_cpu
pycparser=2.22=pyhd8ed1ab_0
pyparsing=3.2.0=pyhd8ed1ab_1
pyside6=6.8.0.2=py310hfd10a26_0
pysocks=1.7.1=pyha2e5f31_6
python=3.10.15=h4a871b0_2_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python-tzdata=2024.2=pyhd8ed1ab_0
python-xxhash=3.5.0=py310ha75aee5_1
python_abi=3.10=5_cp310
pytorch=2.5.1=py3.10_cuda11.8_cudnn9.1.0_0
pytorch-cuda=11.8=h7e8668a_6
pytorch-mutex=1.0=cuda
pytz=2024.1=pyhd8ed1ab_0
pyyaml=6.0.2=py310ha75aee5_1
qhull=2020.2=h434a139_5
qt6-main=6.8.0=h6e8976b_0
re2=2024.07.02=h77b4e00_1
readline=8.2=h8228510_1
regex=2024.11.6=py310ha75aee5_0
requests=2.32.3=pyhd8ed1ab_0
s2n=1.5.7=hd3e8b83_0
safetensors=0.4.5=py310h505e2c1_0
scikit-learn=1.5.2=py310h27f47ee_1
scipy=1.14.1=py310hfcf56fc_1
setuptools=75.3.0=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
snappy=1.2.1=ha2e4443_0
sympy=1.13.3=pyh2585a3b_104
tbb=2021.13.0=h84d6215_0
threadpoolctl=3.5.0=pyhc1e730c_0
tk=8.6.13=noxft_h4845f30_101
tokenizers=0.20.3=py310hb1c6e2c_0
torchtriton=3.1.0=py310
tornado=6.4.1=py310ha75aee5_1
tqdm=4.67.0=pyhd8ed1ab_0
transformers=4.46.2=pyhd8ed1ab_0
typing-extensions=4.12.2=hd8ed1ab_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024b=hc8b5060_0
unicodedata2=15.1.0=py310ha75aee5_1
urllib3=2.2.3=pyhd8ed1ab_0
wayland=1.23.1=h3e06ad9_0
wheel=0.45.0=pyhd8ed1ab_0
xcb-util=0.4.1=hb711507_2
xcb-util-cursor=0.1.5=hb9d3cd8_0
xcb-util-image=0.4.0=hb711507_2
xcb-util-keysyms=0.4.1=hb711507_0
xcb-util-renderutil=0.3.10=hb711507_0
xcb-util-wm=0.4.2=hb711507_0
xkeyboard-config=2.43=hb9d3cd8_0
xorg-libice=1.1.1=hb9d3cd8_1
xorg-libsm=1.2.4=he73a12e_1
xorg-libx11=1.8.10=h4f16b4b_0
xorg-libxau=1.0.11=hb9d3cd8_1
xorg-libxcomposite=0.4.6=hb9d3cd8_2
xorg-libxcursor=1.2.3=hb9d3cd8_0
xorg-libxdamage=1.1.6=hb9d3cd8_0
xorg-libxdmcp=1.1.5=hb9d3cd8_0
xorg-libxext=1.3.6=hb9d3cd8_0
xorg-libxfixes=6.0.1=hb9d3cd8_0
xorg-libxi=1.8.2=hb9d3cd8_0
xorg-libxrandr=1.5.4=hb9d3cd8_0
xorg-libxrender=0.9.11=hb9d3cd8_1
xorg-libxtst=1.2.5=hb9d3cd8_3
xorg-libxxf86vm=1.1.5=hb9d3cd8_4
xorg-xorgproto=2024.1=hb9d3cd8_1
xxhash=0.8.2=hd590300_0
xz=5.2.6=h166bdaf_0
yaml=0.2.5=h7f98852_2
yarl=1.16.0=py310ha75aee5_0
zlib=1.3.1=hb9d3cd8_2
zstandard=0.23.0=py310ha39cb0e_1
zstd=1.5.6=ha6fb4c9_0
